# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Shopify AI Optimizer Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project (one per Shopify store)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectInput"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /projects/{projectId}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project by ID
      parameters:
        - $ref: "#/components/parameters/projectId"
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateProject
      tags: [projects]
      summary: Update project configuration
      parameters:
        - $ref: "#/components/parameters/projectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectInput"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete project
      parameters:
        - $ref: "#/components/parameters/projectId"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /projects/{projectId}/refresh-token:
    post:
      operationId: refreshProjectToken
      tags: [projects]
      summary: Refresh Shopify OAuth access token using client_credentials
      parameters:
        - $ref: "#/components/parameters/projectId"
      responses:
        "200":
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenRefreshResult"
        "400":
          description: Token refresh failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /projects/{projectId}/test-connection:
    post:
      operationId: testProjectConnection
      tags: [projects]
      summary: Test Shopify connection and token validity
      parameters:
        - $ref: "#/components/parameters/projectId"
      responses:
        "200":
          description: Connection test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectionTestResult"

  # ─── PRODUCTS ────────────────────────────────────────────────
...
  /projects/{projectId}/products/{productId}/redesign:
    post:
      operationId: redesignProduct
      tags: [redesign]
      summary: AI redesign a single product with Claude
      parameters:
        - $ref: "#/components/parameters/projectId"
        - $ref: "#/components/parameters/productId"
      responses:
        "200":
          description: Redesigned product data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RedesignResult"
  /projects/{projectId}/products/{productId}/apply-redesign:
    post:
      operationId: applyRedesign
      tags: [redesign]
      summary: Apply redesign to Shopify product
      parameters:
        - $ref: "#/components/parameters/projectId"
        - $ref: "#/components/parameters/productId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplyRedesignInput"
      responses:
        "200":
          description: Apply result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /projects/{projectId}/bulk-redesign:
    post:
      operationId: bulkRedesign
      tags: [redesign]
      summary: Bulk redesign products (all weak or all)
      parameters:
        - $ref: "#/components/parameters/projectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkRedesignInput"
      responses:
        "200":
          description: Bulk redesign job started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkJobResponse"

  # ─── IMAGES ──────────────────────────────────────────────────
...
components:
  parameters:
    projectId:
      name: projectId
      in: path
      required: true
      schema:
        type: integer
    productId:
      name: productId
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success

    # ─── PROJECT ───────────────────────────────────────────
    Project:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        shopDomain:
          type: string
        clientId:
          type: string
        clientSecret:
          type: string
        hasAccessToken:
          type: boolean
        tokenExpiresAt:
          type: ["string", "null"]
        storeNiche:
          type: ["string", "null"]
        brandTone:
          type: ["string", "null"]
        targetAudience:
          type: ["string", "null"]
        storeMarkets:
          type: ["string", "null"]
        replicateApiToken:
          type: ["string", "null"]
        anthropicApiKey:
          type: ["string", "null"]
        autoPilotEnabled:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
        productCount:
          type: ["integer", "null"]
        avgAuditScore:
          type: ["number", "null"]
      required:
        - id
        - name
        - shopDomain
        - clientId
        - clientSecret
        - hasAccessToken
        - autoPilotEnabled
        - createdAt
        - updatedAt
    CreateProjectInput:
      type: object
      properties:
        name:
          type: string
        shopDomain:
          type: string
        clientId:
          type: string
        clientSecret:
          type: string
        storeNiche:
          type: string
        brandTone:
          type: string
        targetAudience:
          type: string
        storeMarkets:
          type: string
        replicateApiToken:
          type: string
        anthropicApiKey:
          type: string
      required:
        - name
        - shopDomain
        - clientId
        - clientSecret
    UpdateProjectInput:
      type: object
      properties:
        name:
          type: string
        shopDomain:
          type: string
        clientId:
          type: string
        clientSecret:
          type: string
        storeNiche:
          type: string
        brandTone:
          type: string
        targetAudience:
          type: string
        storeMarkets:
          type: string
        replicateApiToken:
          type: string
        anthropicApiKey:
          type: string
        autoPilotEnabled:
          type: boolean
    TokenRefreshResult:
      type: object
      properties:
        success:
          type: boolean
        expiresAt:
          type: ["string", "null"]
        message:
          type: string
      required:
        - success
        - message
    ConnectionTestResult:
      type: object
      properties:
        connected:
          type: boolean
        storeName:
          type: ["string", "null"]
        planName:
          type: ["string", "null"]
        productCount:
          type: ["integer", "null"]
        tokenValid:
          type: boolean
        tokenExpiresAt:
          type: ["string", "null"]
        error:
          type: ["string", "null"]
      required:
        - connected
        - tokenValid

    # ─── PRODUCTS ──────────────────────────────────────────
...
    RedesignResult:
      type: object
      properties:
        productId:
          type: string
        title:
          type: string
        bodyHtml:
          type: string
        shortDescription:
          type: string
        price:
          type: string
        compareAtPrice:
          type: string
        tags:
          type: string
        metaTitle:
          type: string
        metaDescription:
          type: string
        photoBrief:
          type: array
          items:
            type: string
        originalTitle:
          type: string
        originalPrice:
          type: string
        priceReasoning:
          type: string
      required:
        - productId
        - title
        - bodyHtml
        - shortDescription
        - price
        - compareAtPrice
        - tags
        - metaTitle
        - metaDescription
        - photoBrief
        - originalTitle
        - originalPrice
    ApplyRedesignInput:
      type: object
      properties:
        redesignId:
          type: integer
        fields:
          type: array
          items:
            type: string
      required:
        - fields
    BulkRedesignInput:
      type: object
      properties:
        mode:
          type: string
        minGrade:
          type: string
      required:
        - mode
    BulkJobResponse:
      type: object
      properties:
        jobId:
          type: string
        status:
          type: string
        totalItems:
          type: integer
        message:
          type: string
      required:
        - jobId
        - status
        - totalItems
        - message

    # ─── IMAGES ────────────────────────────────────────────
...